Helena luz

+++ hackweek-backend/Models/GroupRatingModel.cs
@@ -3,7 +3,7 @@
     public class GroupRatingModel
     {
         public int Id { get; set; }
-        public uint Grade { get; set; }
+        public  double Grade { get; set; }
         public int GroupId { get; set; }
         public int PropositionCriterionId { get; set; }
         public GroupModel? Group { get; set; }
